test(alert): migrate standalone
