+++ core/src/components/alert/test/standalone/alert.e2e.ts
@@ -1,0 +1,18 @@
+import { expect } from '@playwright/test';
+import { configs, test } from '@utils/test/playwright';
+
+test.describe('alert: standalone', () => {
+  configs().forEach(({ config, screenshot, title }) => {
+    test(title('should not have visual regressions'), async ({ page }) => {
+      await page.goto(`/src/components/alert/test/standalone`, config);
+
+      const alert = page.locator('ion-alert');
+      const didPresent = await page.spyOnEvent('ionAlertDidPresent');
+
+      await page.click('#basic');
+      await didPresent.next();
+
+      await expect(alert).toHaveScreenshot(screenshot('alert-standalone'));
+    });
+  });
+});
